# docs/API-CONTRACT.openapi.yaml
openapi: 3.0.3
info:
  title: Nursery Management System API
  version: 0.2.0
  description: |
    Canonical OpenAPI contract for the Nursery Management System API.
    Aligned with current Gin route registration in `api/internal/app/bootstrap/bootstrap.go` and module handlers.
    Includes operational/debug routes and deferred proposed contracts.
  contact:
    name: Nursery API
servers:
  - url: /
    description: Same-origin (dev/proxy)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Authorization: Bearer <access_token>"
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: HttpOnly cookie set after login/refresh
    csrfCookie:
      type: apiKey
      in: cookie
      name: csrf_token
      description: Readable cookie containing CSRF token value
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Must match csrf_token cookie value
    stripeSignature:
      type: apiKey
      in: header
      name: Stripe-Signature
      description: Stripe webhook signature header
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        request_id:
          type: string
          format: uuid
        details:
          description: Optional field-level details
          nullable: true
          type: object
          additionalProperties: true
      example:
        code: validation_error
        message: Invalid request payload.
        request_id: "00000000-0000-0000-0000-000000000001"
        details:
          field: email
...
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in_seconds:
          type: integer
        user:
          $ref: "#/components/schemas/UserResponse"
        active_membership:
          $ref: "#/components/schemas/MembershipResponse"
        available_memberships:
          type: array
          items:
            $ref: "#/components/schemas/MembershipResponse"
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
    MembershipResponse:
      type: object
      properties:
        membership_id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        tenant_name:
          type: string
          description: Display-safe nursery/tenant name for picker labels
        branch_id:
          type: string
          format: uuid
          nullable: true
          description: Null for owner memberships (tenant-wide scope)
        branch_name:
          type: string
          nullable: true
          description: Display-safe branch name for picker labels. Null for owner memberships.
        role:
          type: string
          enum: [owner, manager, practitioner, parent]
    MembershipSelectionRequiredResponse:
      type: object
      description: Pre-session membership selection challenge. No auth cookies, no access token, no session.
      required: [code, message, request_id, available_memberships]
      properties:
        code:
          type: string
          example: membership_selection_required
        message:
          type: string
          description: "Initial: 'Choose a nursery to continue.' Stale: 'That access is no longer available. Choose another nursery or contact your manager.'"
        request_id:
          type: string
          format: uuid
        available_memberships:
          type: array
          items:
            $ref: "#/components/schemas/MembershipResponse"
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: validation_error
            message: Invalid request payload.
            request_id: "00000000-0000-0000-0000-000000000001"
            details:
              field: email
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: unauthorized
            message: Invalid credentials or session.
            request_id: "00000000-0000-0000-0000-000000000001"
...
    ForbiddenScopeSelection:
      description: CSRF failure or scope mismatch
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: forbidden_scope_selection
            message: Access denied.
            request_id: "00000000-0000-0000-0000-000000000001"
...
paths:
...
  /api/v1/auth/login:
    post:
      operationId: authLogin
      summary: Login with email/password
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                membership_id:
                  type: string
                  format: uuid
                  description: Required after a membership_selection_required challenge, or when the client already knows the intended membership. Omit for initial sign-in.
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
          headers:
            Set-Cookie:
              description: "Sets refresh_token (HttpOnly) and csrf_token cookies"
              schema:
                type: string
        "400":
          description: Validation error or membership selection challenge
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErrorResponse"
                  - $ref: "#/components/schemas/MembershipSelectionRequiredResponse"
              examples:
                validation_error:
                  value:
                    code: validation_error
                    message: Invalid request payload.
                    request_id: "00000000-0000-0000-0000-000000000001"
                    details:
                      field: membership_id
                membership_selection_required:
                  value:
                    code: membership_selection_required
                    message: "Choose a nursery to continue."
                    request_id: "00000000-0000-0000-0000-000000000001"
                    available_memberships:
                      - membership_id: "00000000-0000-0000-0000-000000000010"
                        tenant_id: "00000000-0000-0000-0000-000000000020"
                        tenant_name: "Little Sprouts Nursery"
                        branch_id: "00000000-0000-0000-0000-000000000030"
                        branch_name: "Main Branch"
                        role: manager
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/refresh:
    post:
      operationId: authRefresh
      summary: Refresh access token
      tags: [Auth]
      description: Cookie-backed, CSRF-protected session action. No request body.
      security:
        - refreshTokenCookie: []
          csrfCookie: []
          csrfHeader: []
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenScopeSelection"
  /api/v1/auth/logout:
    post:
      operationId: authLogout
      summary: Logout and clear session
      tags: [Auth]
      description: Cookie-backed. CSRF-protected when refresh cookie present. Clears cookies. Idempotent.
      security:
        - refreshTokenCookie: []
          csrfCookie: []
          csrfHeader: []
        - {}
      responses:
        "204":
          description: Session cleared
  /api/v1/auth/switch-membership:
    post:
      operationId: authSwitchMembership
      summary: Switch active membership
      tags: [Auth]
      description: Cookie-backed, CSRF-protected session action.
      security:
        - refreshTokenCookie: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [membership_id]
              properties:
                membership_id:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Membership switched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenScopeSelection"
...